Reworked puck collision system
Reworked puck collision detection system to employ better detection that checks for intersections between the puck's path and each wall rather than checking if the puck passed the inner walls of the paddles. This was unreliable and resulted in the system detecting a collision if the puck moved behind the wall from above or below. It needed to detect if it moved through the wall, not around, thus a line intersection detection system was implemented.